Factor out hp-action-tenant-children.tsx.
